+++ packages/expo-camera/android/src/main/java/expo/modules/camera/CameraViewHelper.java
@@ -7,7 +7,6 @@ import android.graphics.Paint;
 import android.media.CamcorderProfile;
 import android.os.Bundle;
 import androidx.exifinterface.media.ExifInterface;
-import android.os.Build;
 import android.view.ViewGroup;
 
 import com.google.android.cameraview.CameraView;
@@ -19,7 +18,6 @@ import java.util.List;
 import java.util.Locale;
 
 import org.unimodules.core.interfaces.services.EventEmitter;
-import org.unimodules.interfaces.barcodescanner.BarCodeScannerResult;
 import org.unimodules.interfaces.facedetector.FaceDetector;
 import expo.modules.camera.events.BarCodeScannedEvent;
 import expo.modules.camera.events.CameraMountErrorEvent;
@@ -27,6 +25,7 @@ import expo.modules.camera.events.CameraReadyEvent;
 import expo.modules.camera.events.FaceDetectionErrorEvent;
 import expo.modules.camera.events.FacesDetectedEvent;
 import expo.modules.camera.events.PictureSavedEvent;
+import expo.modules.interfaces.barcodescanner.BarCodeScannerResult;
 
 public class CameraViewHelper {
   // Mount error event