+++ packages/expo-barcode-scanner/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Enable kotlin in all modules. ([#12716](https://github.com/expo/expo/pull/12716) by [@wschurman](https://github.com/wschurman))
 
+### 💡 Others
+
+- Migrated module interface from `unimodules-barcode-scanner-interface` to `expo-modules-core`. ([#12912](https://github.com/expo/expo/pull/12912) by [@tsapeta](https://github.com/tsapeta))
+
 ## 10.1.2 — 2021-04-13
 
 _This version does not introduce any user-facing changes._