+++ packages/react-native-unimodules/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### 💡 Others
 
-- Migrated camera, constants and sensors interfaces from their own packages to `expo-modules-core`. ([#12868](https://github.com/expo/expo/pull/12868), [#12876](https://github.com/expo/expo/pull/12876), [#12888](https://github.com/expo/expo/pull/12888) by [@tsapeta](https://github.com/tsapeta))
+- Migrated camera, constants, sensors and barcode scanner interfaces from their own packages to `expo-modules-core`. ([#12868](https://github.com/expo/expo/pull/12868), [#12876](https://github.com/expo/expo/pull/12876), [#12888](https://github.com/expo/expo/pull/12888), [#12912](https://github.com/expo/expo/pull/12912) by [@tsapeta](https://github.com/tsapeta))
 
 ## 0.13.3 — 2021-04-13